Backend/Tree/requirements.txt
asgiref==3.7.2
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
contourpy==1.2.1
crashtest==0.4.1
cycler==0.12.1
distlib==0.3.8
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.3
fonttools==4.51.0
fsspec==2024.3.1
idna==3.6
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.3
keyring==24.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry-core==1.9.0
poetry-plugin-export==1.7.1
psutil==5.9.8
ptyprocess==0.7.0
py-cpuinfo==9.0.0
PyJWT==2.8.0
pyparsing==3.1.2
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.8.0
requests==2.31.0
requests-toolbelt==1.0.0
scipy==1.13.0
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
sympy==1.12
thop==0.1.1.post2209072238
tomlkit==0.12.4
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
trove-classifiers==2024.3.25
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.1.44
urllib3==2.2.1
virtualenv==20.25.1
